Added support for VS 22
