+++ TotalCommanderLauncher.Shared/ActivePanelConverter.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
+{
+    public class ActivePanelConverter : EnumConverter<ActivePanel>
+    { }
+}
