+++ TotalCommanderLauncher.Shared/Images/Icon.pdn
