+++ TotalCommanderLauncher.Shared/Images/Icon.png
