+++ TotalCommanderLauncher.Shared/Key.snk
