+++ TotalCommanderLauncher.Shared/Licenses/MS-PL.txt
