+++ TotalCommanderLauncher.Shared/Resources/Images.pdn
