+++ TotalCommanderLauncher.Shared/Resources/Images.png
