+++ TotalCommanderLauncher.Shared/Resources/Package.ico
