--- TotalCommanderLauncher/ActivePanel.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
-{
-    public enum ActivePanel
-    {
-        [Description("Left Panel")]
-        Left,
-
-        [Description("Right Panel")]
-        Right
-    }
-}
