--- TotalCommanderLauncher/ActivePanelConverter.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
-{
-    public class ActivePanelConverter : EnumConverter<ActivePanel>
-    { }
-}
