--- TotalCommanderLauncher/CommandIds.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
-{
-    internal static class CommandIds
-    {
-        internal const uint TotalCommander = 0x201;
-    }
-}
