--- TotalCommanderLauncher/GlobalSuppressions.cs
@@ -1,1 +1,0 @@
-﻿[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Design", "CA1017:MarkAssembliesWithComVisible")]
