--- TotalCommanderLauncher/Guids.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
-{
-    internal static class Guids
-    {
-        internal const string Package = "504f9f56-9629-48c7-adb9-cbcce53bf9b7";
-        internal const string Options = "10a4a8b0-08aa-4b25-a95e-b95a03448a1c";
-        internal static readonly Guid MenuGroup = new Guid("40cf569c-214e-41db-b92d-b8066cf86495");
-    }
-}
