--- TotalCommanderLauncher/NativeMethods.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GrzegorzKozub.VisualStudioExtensions.TotalCommanderLauncher
-{
-    internal static class NativeMethods
-    {
-        [DllImport("user32.dll")]
-        internal static extern bool SetForegroundWindow(IntPtr hWnd);
-    }
-}
