--- TotalCommanderLauncher/packages.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EnvDTE" version="8.0.2" targetFramework="net45" />
-  <package id="EnvDTE100" version="10.0.3" targetFramework="net45" />
-  <package id="EnvDTE80" version="8.0.3" targetFramework="net45" />
-  <package id="EnvDTE90" version="9.0.3" targetFramework="net45" />
-  <package id="EnvDTE90a" version="9.0.3" targetFramework="net45" />
-  <package id="Microsoft.VisualStudio.CommandBars" version="8.0.0.1" targetFramework="net45" />
-  <package id="Microsoft.VisualStudio.Sdk.BuildTasks.14.0" version="14.0.23-pre" targetFramework="net45" developmentDependency="true" />
-  <package id="Microsoft.VisualStudio.SDK.EmbedInteropTypes" version="15.0.16" targetFramework="net45" />
-  <package id="stdole" version="7.0.3302" targetFramework="net45" />
-  <package id="VSSDK.DTE" version="7.0.4" targetFramework="net45" />
-  <package id="VSSDK.GraphModel" version="11.0.4" targetFramework="net45" />
-  <package id="VSSDK.IDE" version="7.0.4" targetFramework="net45" />
-  <package id="VSSDK.IDE.10" version="10.0.4" targetFramework="net45" />
-  <package id="VSSDK.IDE.11" version="11.0.4" targetFramework="net45" />
-  <package id="VSSDK.IDE.8" version="8.0.4" targetFramework="net45" />
-  <package id="VSSDK.IDE.9" version="9.0.4" targetFramework="net45" />
-  <package id="VSSDK.OLE.Interop" version="7.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.11" version="11.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Immutable.10" version="10.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Immutable.11" version="11.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Interop" version="7.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Interop.10" version="10.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Interop.11" version="11.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Interop.8" version="8.0.4" targetFramework="net45" />
-  <package id="VSSDK.Shell.Interop.9" version="9.0.4" targetFramework="net45" />
-  <package id="VSSDK.TextManager.Interop" version="7.0.4" targetFramework="net45" />
-  <package id="VSSDK.TextManager.Interop.8" version="8.0.4" targetFramework="net45" />
-</packages>
